Refactor: 채팅 전체 코드 리팩토링 및 쿼리 개선

+++ src/main/java/com/dataury/soloJ/domain/chat/dto/ChatRoomRequestDto.java
@@ -15,7 +15,7 @@ public class ChatRoomRequestDto {
         private String description;
         private Long contentId; // TouristSpot의 contentId
         private LocalDateTime joinDate;
-        private int maxMembers; // 최대 인원수
+        private Long maxMembers; // 최대 인원수
     }
 
     @Data
